refactor: Simplify search components and improve code organization
